clean up unit test
