+++ nnd-data-poll-service/src/main/java/gov/cdc/nnddatapollservice/srte/dao/SrteDataPersistentDAO.java
@@ -1,33 +1,14 @@
 package gov.cdc.nnddatapollservice.srte.dao;
 
-import com.google.gson.FieldNamingPolicy;
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.reflect.TypeToken;
 import gov.cdc.nnddatapollservice.rdb.dto.PollDataSyncConfig;
-import gov.cdc.nnddatapollservice.repository.srte.CodeToConditionRepository;
-import gov.cdc.nnddatapollservice.repository.srte.model.CodeToCondition;
 import gov.cdc.nnddatapollservice.service.model.LogResponseModel;
-import gov.cdc.nnddatapollservice.share.HandleError;
 import gov.cdc.nnddatapollservice.share.JdbcTemplateUtil;
-import gov.cdc.nnddatapollservice.share.PollServiceUtil;
-import gov.cdc.nnddatapollservice.srte.dto.CodeToConditionDto;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
-import org.springframework.jdbc.core.namedparam.SqlParameterSourceUtils;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Service;
-
-import java.lang.reflect.Type;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
 
 import static gov.cdc.nnddatapollservice.constant.ConstantValue.*;
 import static gov.cdc.nnddatapollservice.share.StringUtil.getStackTraceAsString;
