+++ nnd-data-poll-service/src/main/java/gov/cdc/nnddatapollservice/universal/service/UniversalDataHandlingService.java
@@ -3,11 +3,11 @@ package gov.cdc.nnddatapollservice.universal.service;
 import gov.cdc.nnddatapollservice.edx_nbs_odse.dao.EdxNbsOdseDataPersistentDAO;
 import gov.cdc.nnddatapollservice.exception.DataPollException;
 import gov.cdc.nnddatapollservice.rdb.dto.PollDataSyncConfig;
-import gov.cdc.nnddatapollservice.universal.dao.UniversalDataPersistentDAO;
-import gov.cdc.nnddatapollservice.universal.service.interfaces.IUniversalDataHandlingService;
 import gov.cdc.nnddatapollservice.service.interfaces.IPollCommonService;
 import gov.cdc.nnddatapollservice.service.interfaces.IS3DataService;
 import gov.cdc.nnddatapollservice.service.model.LogResponseModel;
+import gov.cdc.nnddatapollservice.universal.dao.UniversalDataPersistentDAO;
+import gov.cdc.nnddatapollservice.universal.service.interfaces.IUniversalDataHandlingService;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
