+++ nnd-data-poll-service/src/test/java/gov/cdc/nnddatapollservice/universal/dao/UniversalDataPersistentDAOTest.java
@@ -2,14 +2,17 @@ package gov.cdc.nnddatapollservice.universal.dao;
 
 import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
-import gov.cdc.nnddatapollservice.universal.dto.NrtObservationCodedDto;
-import gov.cdc.nnddatapollservice.universal.dto.NrtObservationDto;
+import gov.cdc.nnddatapollservice.rdb.dto.PollDataSyncConfig;
 import gov.cdc.nnddatapollservice.repository.rdb_modern.NrtObservationCodedRepository;
 import gov.cdc.nnddatapollservice.repository.rdb_modern.NrtObservationRepository;
 import gov.cdc.nnddatapollservice.repository.rdb_modern.model.NrtObservation;
 import gov.cdc.nnddatapollservice.repository.rdb_modern.model.NrtObservationCoded;
+import gov.cdc.nnddatapollservice.service.model.LogResponseModel;
 import gov.cdc.nnddatapollservice.share.HandleError;
+import gov.cdc.nnddatapollservice.share.JdbcTemplateUtil;
 import gov.cdc.nnddatapollservice.share.PollServiceUtil;
+import gov.cdc.nnddatapollservice.universal.dto.NrtObservationCodedDto;
+import gov.cdc.nnddatapollservice.universal.dto.NrtObservationDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
@@ -21,6 +24,7 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 
 import javax.sql.DataSource;
 import java.lang.reflect.Type;
+import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
@@ -33,7 +37,7 @@ import static org.mockito.Mockito.*;
 
 class UniversalDataPersistentDAOTest {
     @Mock
-    private JdbcTemplate jdbcTemplate;
+    private JdbcTemplateUtil jdbcTemplate;
     @Mock
     private DataSource dataSource;
     @Mock
@@ -50,175 +54,26 @@ class UniversalDataPersistentDAOTest {
     @BeforeEach
     void setUp() {
         MockitoAnnotations.openMocks(this);
-        when(jdbcTemplate.getDataSource()).thenReturn(dataSource);
         universalDataPersistentDAO.batchSize = 10000;
     }
 
-    @Test
-    void testPersistingGenericTable_ExceptionCase_WritesToFile()  {
-        // Arrange
-        String tableName = "GENERIC_TABLE";
-        String jsonData = "[{\"key\": \"value\"}]";
-        SimpleJdbcInsert mockSimpleJdbcInsert = mock(SimpleJdbcInsert.class);
-
-        doThrow(new RuntimeException("SQL Error")).when(mockSimpleJdbcInsert).executeBatch(any(SqlParameterSource[].class));
-
-        StringBuilder logBuilder = new StringBuilder();
-
-        StringBuilder resultLog = universalDataPersistentDAO.persistingGenericTable(logBuilder, tableName, jsonData,
-                "key", true);
-        assertNotNull(resultLog.toString());
-    }
 
     @Test
     void saveRdbModernData(){
         String jsondata = "[{\"CONFIRMATION_METHOD_KEY\":1,\"CONFIRMATION_METHOD_CD\":null,\"CONFIRMATION_METHOD_DESC\":null},\n" +
                 "{\"CONFIRMATION_METHOD_KEY\":23,\"CONFIRMATION_METHOD_CD\":\"MR\",\"CONFIRMATION_METHOD_DESC\":\"Medical record review\"}]";
-        var recordsSaved = universalDataPersistentDAO.saveRdbModernData("TEST_TABLE", jsondata,
-                "key", true);
+        PollDataSyncConfig config = new PollDataSyncConfig();
+        config.setTableName("TEST_TABLE");
+        config.setKeyList("key");
+
+
+        when(jdbcTemplate.persistingGenericTable(
+                jsondata,
+                config, true)).thenReturn(new LogResponseModel());
+        var recordsSaved = universalDataPersistentDAO.saveRdbModernData(config, jsondata, true);
+
+
         assertNotNull(recordsSaved);
-    }
-
-    @Test
-    void testSaveRdbModernData_withValidData() {
-        String tableName = "test_table";
-        String jsonData = "[{\"key1\":\"value1\", \"key2\":\"value2\"}]";
-
-        List<Map<String, Object>> records = PollServiceUtil.jsonToListOfMap(jsonData);
-        assert records != null;
-
-        SimpleJdbcInsert mockSimpleJdbcInsert = mock(SimpleJdbcInsert.class);
-        when(mockSimpleJdbcInsert.withTableName(anyString())).thenReturn(mockSimpleJdbcInsert);
-        when(mockSimpleJdbcInsert.executeBatch(any(SqlParameterSource[].class))).thenReturn(new int[]{1});
-
-        doNothing().when(jdbcTemplate).execute(anyString());
-
-        var noOfRecordsSaved = universalDataPersistentDAO.saveRdbModernData(tableName, jsonData,
-                "key", true);
-
-        assertNotNull( noOfRecordsSaved);
-    }
-
-    @Test
-    void testSaveRdbModernData_withNoData() {
-        String tableName = "test_table";
-        String jsonData = "[]";
-
-        SimpleJdbcInsert mockSimpleJdbcInsert = mock(SimpleJdbcInsert.class);
-        when(mockSimpleJdbcInsert.withTableName(anyString())).thenReturn(mockSimpleJdbcInsert);
-
-        doNothing().when(jdbcTemplate).execute(anyString());
-
-        var noOfRecordsSaved = universalDataPersistentDAO.saveRdbModernData(tableName, jsonData,
-                "key", true);
-
-        assertEquals("SUCCESS", noOfRecordsSaved);
-    }
-
-    @Test
-    void testDeleteTable() {
-        String tableName = "test_table";
-
-        doNothing().when(jdbcTemplate).execute(anyString());
-
-        universalDataPersistentDAO.deleteTable(tableName);
-
-        verify(jdbcTemplate, times(1)).execute("delete FROM " + tableName);
-    }
-    @Test
-    void deleteTable_shouldLogErrorOnException() {
-        String tableName = "non_existing_table";
-        doThrow(new RuntimeException("Simulated exception")).when(jdbcTemplate).execute(anyString());
-        universalDataPersistentDAO.deleteTable(tableName);
-        verify(jdbcTemplate).execute("delete FROM " + tableName);
-    }
-
-    @Test
-    void testSaveRdbModernData_NrtObservation_Success() {
-        // Arrange
-        String tableName = "NRT_OBSERVATION";
-        String jsonData = "[{\"observation_uid\": 1, \"class_cd\": \"OBS\"}]";
-
-        Type resultType = new TypeToken<List<NrtObservationDto>>() {}.getType();
-        List<NrtObservationDto> observationDtos = new ArrayList<>();
-        observationDtos.add(new NrtObservationDto());
-
-        when(gson.fromJson(jsonData, resultType)).thenReturn(observationDtos);
-
-        // Act
-        universalDataPersistentDAO.saveRdbModernData(tableName, jsonData,
-                "key", true);
-
-        // Assert
-        verify(nrtObservationRepository, times(1)).save(any(NrtObservation.class));
-    }
-
-    @Test
-    void testSaveRdbModernData_NrtObservationCoded_Success() {
-        // Arrange
-        String tableName = "nrt_observation_coded";
-        String jsonData = "[{\"observation_uid\": 1, \"class_cd\": \"OBS\"}]";
-
-        Type resultType = new TypeToken<List<NrtObservationCodedDto>>() {}.getType();
-        List<NrtObservationCodedDto> observationDtos = new ArrayList<>();
-        observationDtos.add(new NrtObservationCodedDto());
-
-        when(gson.fromJson(jsonData, resultType)).thenReturn(observationDtos);
-
-        // Act
-        universalDataPersistentDAO.saveRdbModernData(tableName, jsonData,
-                "key", true);
-
-        // Assert
-        verify(nrtObservationCodedRepository, times(1)).save(any(NrtObservationCoded.class));
-    }
-
-
-    @Test
-    void testSaveRdbModernData_Else_Success() {
-        // Arrange
-        String tableName = "SOME_TABLE";
-        String jsonData = "[{\"key1\": \"value1\", \"key2\": \"value2\"}]";
-
-        List<Map<String, Object>> records = PollServiceUtil.jsonToListOfMap(jsonData);
-        assert records != null;
-
-        SimpleJdbcInsert mockSimpleJdbcInsert = mock(SimpleJdbcInsert.class);
-        when(mockSimpleJdbcInsert.withTableName(anyString())).thenReturn(mockSimpleJdbcInsert);
-        when(mockSimpleJdbcInsert.executeBatch(any(SqlParameterSource[].class))).thenReturn(new int[]{1});
-
-        doNothing().when(jdbcTemplate).execute(anyString());
-
-        // Act
-        String result = universalDataPersistentDAO.saveRdbModernData(tableName, jsonData,
-                "key", true);
-
-        // Assert
-        assertNotNull(result);
-    }
-
-    @Test
-    void testSaveRdbModernData_Else_Success_2() {
-        // Arrange
-        universalDataPersistentDAO.batchSize = 0;
-        String tableName = "SOME_TABLE";
-        String jsonData = "[{\"key1\": \"value1\", \"key2\": \"value2\"}]";
-
-        List<Map<String, Object>> records = PollServiceUtil.jsonToListOfMap(jsonData);
-        assert records != null;
-
-        SimpleJdbcInsert mockSimpleJdbcInsert = mock(SimpleJdbcInsert.class);
-        when(mockSimpleJdbcInsert.withTableName(anyString())).thenReturn(mockSimpleJdbcInsert);
-        when(mockSimpleJdbcInsert.executeBatch(any(SqlParameterSource[].class))).thenReturn(new int[]{1});
-
-        doNothing().when(jdbcTemplate).execute(anyString());
-
-        // Act
-        String result = universalDataPersistentDAO.saveRdbModernData(tableName, jsonData,
-                "key", true);
-
-        // Assert
-        assertNotNull(result);
     }
 
 }
