+++ nnd-data-poll-service/src/test/java/gov/cdc/nnddatapollservice/universal/service/RdbModernDataHandlingServiceTest.java
@@ -2,10 +2,11 @@ package gov.cdc.nnddatapollservice.universal.service;
 
 import gov.cdc.nnddatapollservice.exception.DataPollException;
 import gov.cdc.nnddatapollservice.rdb.dto.PollDataSyncConfig;
-import gov.cdc.nnddatapollservice.universal.dao.UniversalDataPersistentDAO;
 import gov.cdc.nnddatapollservice.service.interfaces.IPollCommonService;
 import gov.cdc.nnddatapollservice.service.interfaces.IS3DataService;
+import gov.cdc.nnddatapollservice.service.model.LogResponseModel;
 import gov.cdc.nnddatapollservice.share.TimestampUtil;
+import gov.cdc.nnddatapollservice.universal.dao.UniversalDataPersistentDAO;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
@@ -16,7 +17,7 @@ import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.List;
 
-import static gov.cdc.nnddatapollservice.constant.ConstantValue.*;
+import static gov.cdc.nnddatapollservice.constant.ConstantValue.RDB;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.*;
 
@@ -58,7 +59,7 @@ class RdbModernDataHandlingServiceTest {
         when(iPollCommonService.checkPollingIsInitailLoad(configTableList)).thenReturn(true);
 
         universalDataHandlingService.handlingExchangedData("RDB");
-        verify(universalDataPersistentDAO, times(1)).deleteTable(anyString());
+        verify(iPollCommonService, times(1)).deleteTable(anyString());
     }
 
     @Test
@@ -78,7 +79,7 @@ class RdbModernDataHandlingServiceTest {
         when(iPollCommonService.checkPollingIsInitailLoad(configTableList)).thenReturn(false);
 
         universalDataHandlingService.handlingExchangedData("RDB");
-        verify(universalDataPersistentDAO, times(0)).deleteTable(anyString());
+        verify(iPollCommonService, times(0)).deleteTable(anyString());
     }
 
     @Test
@@ -88,13 +89,18 @@ class RdbModernDataHandlingServiceTest {
         String tableName = "exampleTable";
         universalDataHandlingService.storeJsonInLocalFolder= true;
         when(iPollCommonService.callDataCountEndpoint(anyString(), anyBoolean(), anyString())).thenReturn(1000);
+
+        PollDataSyncConfig config = new PollDataSyncConfig();
+        config.setTableName(tableName);
+        config.setKeyList("key");
+        config.setSourceDb("RDB");
         // Act
-        universalDataHandlingService.pollAndPersistRDBMOdernData("RDB", tableName, true, "key", true);
+        universalDataHandlingService.pollAndPersistData(true, config);
 
         // Assert
         verify(iPollCommonService, times(1)).writeJsonDataToFile(anyString(), anyString(),
                 any(),anyString());
-        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLogLocalDir(anyString(), any(), anyString());
+//        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLogLocalDir(anyString(), any(), any());
     }
 
     private void setupServiceWithMockedDependencies() throws DataPollException {
@@ -102,7 +108,7 @@ class RdbModernDataHandlingServiceTest {
         when(iPollCommonService.decodeAndDecompress(anyString())).thenReturn("{\"data\": \"example\"}");
         when(iPollCommonService.getCurrentTimestamp()).thenReturn("2023-01-01T00:00:00Z");
         when(iPollCommonService.getLastUpdatedTime(anyString())).thenReturn("2023-01-01T00:00:00Z");
-        when(iPollCommonService.callDataExchangeEndpoint(anyString(), anyBoolean(), anyString(), anyBoolean(), anyString(), anyString())).thenReturn("encodedData");
+        when(iPollCommonService.callDataExchangeEndpoint(anyString(), anyBoolean(), anyString(), anyBoolean(), anyString(), anyString(), anyBoolean())).thenReturn("encodedData");
 
 
 
@@ -117,8 +123,12 @@ class RdbModernDataHandlingServiceTest {
         when(iPollCommonService.callDataCountEndpoint(anyString(), anyBoolean(), anyString()))
                 .thenThrow(new RuntimeException(expectedErrorMessage));
 
+        PollDataSyncConfig config = new PollDataSyncConfig();
+        config.setTableName(tableName);
+        config.setKeyList("key");
+        config.setSourceDb("RDB");
         // Act
-        universalDataHandlingService.pollAndPersistRDBMOdernData("RDB", tableName, true, "key", true);
+        universalDataHandlingService.pollAndPersistData( true, config);
         // Assert
         verify(iPollCommonService, never()).updateLastUpdatedTimeAndLog(eq(tableName), any(), any());
     }
@@ -131,11 +141,15 @@ class RdbModernDataHandlingServiceTest {
         when(iPollCommonService.getCurrentTimestamp()).thenReturn("2024-09-17T00:00:00Z");
         when(iPollCommonService.callDataCountEndpoint(anyString(), anyBoolean(), anyString()))
                 .thenReturn(1000);
-        when(iPollCommonService.callDataExchangeEndpoint(anyString(), anyBoolean(), anyString(), anyBoolean(), anyString(), anyString()))
+        when(iPollCommonService.callDataExchangeEndpoint(anyString(), anyBoolean(), anyString(), anyBoolean(), anyString(), anyString(), anyBoolean()))
                 .thenThrow(new RuntimeException(expectedErrorMessage));
 
+        PollDataSyncConfig config = new PollDataSyncConfig();
+        config.setTableName(tableName);
+        config.setKeyList("key");
+        config.setSourceDb("RDB");
         // Act
-        universalDataHandlingService.pollAndPersistRDBMOdernData("RDB", tableName, true, "key", true);
+        universalDataHandlingService.pollAndPersistData( true, config);
 
         // Assert
         verify(iPollCommonService, never()).updateLastUpdatedTimeAndLog(eq(tableName), any(), any());
@@ -155,96 +169,21 @@ class RdbModernDataHandlingServiceTest {
                 anyString(),
                 eq(false),
                 anyString(),
-                anyString()))
+                anyString(),
+                anyBoolean()))
                 .thenThrow(new RuntimeException(expectedErrorMessage));
 
+        PollDataSyncConfig config = new PollDataSyncConfig();
+        config.setTableName(tableName);
+        config.setKeyList("key");
+        config.setSourceDb("RDB");
         // Act
-        universalDataHandlingService.pollAndPersistRDBMOdernData("RDB", tableName, true, "key", true);
+        universalDataHandlingService.pollAndPersistData(true, config);
 
         // Assert
         verify(iPollCommonService, never()).updateLastUpdatedTimeAndLog(eq(tableName), any(), any());
     }
 
-
-
-    @Test
-    void testPollAndPersistRDBData_TimeStampSelection_StoreInSql() throws DataPollException {
-        // Arrange
-        String tableName = "NRT_OBSERVATION";
-        boolean isInitialLoad = false;
-
-        // Mocking common service methods
-        when(iPollCommonService.getLastUpdatedTime(anyString())).thenReturn("2024-10-01 12:00:00");
-        when(iPollCommonService.callDataCountEndpoint(anyString(), anyBoolean(), anyString())).thenReturn(100);
-        when(iPollCommonService.callDataExchangeEndpoint(anyString(), anyBoolean(), anyString(), anyBoolean(), anyString(), anyString())).thenReturn("mockEncodedData");
-        when(iPollCommonService.decodeAndDecompress(anyString())).thenReturn("mockRawJsonData");
-
-        // Setup universalDataHandlingService flags
-        universalDataHandlingService.storeInSql = true;
-        universalDataHandlingService.storeJsonInS3 = false;
-        universalDataHandlingService.storeJsonInLocalFolder = false;
-
-        // Act
-        universalDataHandlingService.pollAndPersistRDBMOdernData("RDB", tableName, isInitialLoad, "key", true);
-
-        // Assert
-        verify(iPollCommonService, times(1)).getLastUpdatedTime(tableName);
-        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLog(eq(tableName), any(Timestamp.class), anyString());
-    }
-
-
-    @Test
-    void testPollAndPersistRDBData_TimeStampSelection_StoreJsonInS3() throws DataPollException {
-        // Arrange
-        String tableName = "NRT_OBSERVATION";
-        boolean isInitialLoad = false;
-
-        // Mocking common service methods
-        when(iPollCommonService.getLastUpdatedTimeS3(anyString())).thenReturn("2024-10-01 12:00:00");
-        when(iPollCommonService.callDataCountEndpoint(anyString(), anyBoolean(), anyString())).thenReturn(100);
-        when(iPollCommonService.callDataExchangeEndpoint(anyString(), anyBoolean(), anyString(), anyBoolean(), anyString(), anyString())).thenReturn("mockEncodedData");
-        when(iPollCommonService.decodeAndDecompress(anyString())).thenReturn("mockRawJsonData");
-
-        // Setup universalDataHandlingService flags
-        universalDataHandlingService.storeInSql = false;
-        universalDataHandlingService.storeJsonInS3 = true;
-        universalDataHandlingService.storeJsonInLocalFolder = false;
-
-        // Act
-        universalDataHandlingService.pollAndPersistRDBMOdernData("RDB", tableName, isInitialLoad, "key", true);
-
-        // Assert
-        verify(iPollCommonService, times(1)).getLastUpdatedTimeS3(tableName);
-        verify(is3DataService, times(1)).persistToS3MultiPart(eq(RDB), anyString(), eq(tableName), any(Timestamp.class), eq(isInitialLoad));
-        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLogS3(eq(tableName), any(Timestamp.class), anyString());
-    }
-
-
-    @Test
-    void testPollAndPersistRDBData_TimeStampSelection_StoreJsonInLocalFolder() throws DataPollException {
-        // Arrange
-        String tableName = "NRT_OBSERVATION";
-        boolean isInitialLoad = false;
-
-        // Mocking common service methods
-        when(iPollCommonService.getLastUpdatedTimeLocalDir(anyString())).thenReturn("2024-10-01 12:00:00");
-        when(iPollCommonService.callDataCountEndpoint(anyString(), anyBoolean(), anyString())).thenReturn(100);
-        when(iPollCommonService.callDataExchangeEndpoint(anyString(), anyBoolean(), anyString(), anyBoolean(), anyString(), anyString())).thenReturn("mockEncodedData");
-        when(iPollCommonService.decodeAndDecompress(anyString())).thenReturn("mockRawJsonData");
-
-        // Setup universalDataHandlingService flags
-        universalDataHandlingService.storeInSql = false;
-        universalDataHandlingService.storeJsonInS3 = false;
-        universalDataHandlingService.storeJsonInLocalFolder = true;
-
-        // Act
-        universalDataHandlingService.pollAndPersistRDBMOdernData("RDB", tableName, isInitialLoad, "key", true);
-
-        // Assert
-        verify(iPollCommonService, times(1)).getLastUpdatedTimeLocalDir(tableName);
-        verify(iPollCommonService, times(1)).writeJsonDataToFile(eq(RDB), eq(tableName), any(Timestamp.class), anyString());
-        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLogLocalDir(eq(tableName), any(Timestamp.class), anyString());
-    }
 
 
     @Test
@@ -260,12 +199,18 @@ class RdbModernDataHandlingServiceTest {
         universalDataHandlingService.storeInSql = true;
         universalDataHandlingService.storeJsonInS3 = false;
 
+        PollDataSyncConfig config = new PollDataSyncConfig();
+        config.setTableName(tableName);
+        config.setKeyList("key");
+        config.setSourceDb("RDB");
+
         // Act
-        universalDataHandlingService.updateDataHelper(exceptionAtApiLevel, tableName, timestamp,
-                rawJsonData, isInitialLoad, log, "RDB", "key");
+        universalDataHandlingService.updateDataHelper(exceptionAtApiLevel, timestamp,
+                rawJsonData, isInitialLoad, log, timestamp, config);
 
         // Assert
-        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLog(tableName, timestamp, API_LEVEL + log);
+        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLog(eq(tableName),
+                eq(timestamp), any());
         verifyNoMoreInteractions(iPollCommonService);
     }
 
@@ -282,13 +227,17 @@ class RdbModernDataHandlingServiceTest {
 
         universalDataHandlingService.storeInSql = false;
         universalDataHandlingService.storeJsonInS3 = false;
-
+        PollDataSyncConfig config = new PollDataSyncConfig();
+        config.setTableName(tableName);
+        config.setKeyList("key");
+        config.setSourceDb("RDB");
         // Act
-        universalDataHandlingService.updateDataHelper(exceptionAtApiLevel, tableName, timestamp,
-                rawJsonData, isInitialLoad, log, "RDB", "key");
+        universalDataHandlingService.updateDataHelper(exceptionAtApiLevel, timestamp,
+                rawJsonData, isInitialLoad, log, timestamp, config);
 
         // Assert
-        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLogLocalDir(tableName, timestamp, API_LEVEL + log);
+        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLogLocalDir(
+                eq(tableName), eq(timestamp), any());
         verifyNoMoreInteractions(iPollCommonService);
     }
 
@@ -308,15 +257,21 @@ class RdbModernDataHandlingServiceTest {
         universalDataHandlingService.storeJsonInS3 = true;
 
         when(is3DataService.persistToS3MultiPart(anyString(), anyString(), anyString(), any(Timestamp.class), anyBoolean()))
-                .thenReturn("S3 Save Success");
+                .thenReturn(new LogResponseModel());
 
+        PollDataSyncConfig config = new PollDataSyncConfig();
+        config.setTableName(tableName);
+        config.setKeyList("key");
+        config.setSourceDb("RDB");
         // Act
-        universalDataHandlingService.updateDataHelper(exceptionAtApiLevel, tableName,
-                timestamp, rawJsonData, isInitialLoad, log, "RDB", "key");
+        universalDataHandlingService.updateDataHelper(exceptionAtApiLevel,
+                timestamp, rawJsonData, isInitialLoad, log, timestamp, config);
 
         // Assert
-        verify(is3DataService, times(1)).persistToS3MultiPart(RDB, rawJsonData, tableName, timestamp, isInitialLoad);
-        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLogS3(tableName, timestamp, S3_LOG + "S3 Save Success");
+        verify(is3DataService, times(1)).persistToS3MultiPart(
+                RDB, rawJsonData, tableName, timestamp, isInitialLoad);
+        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLogS3(
+                eq(tableName), eq(timestamp), any());
     }
 
 
@@ -333,15 +288,20 @@ class RdbModernDataHandlingServiceTest {
         universalDataHandlingService.storeInSql = false;
         universalDataHandlingService.storeJsonInS3 = false;
 
-        when(iPollCommonService.writeJsonDataToFile(anyString(), anyString(), any(Timestamp.class), anyString())).thenReturn("Local File Save Success");
-
+        when(iPollCommonService.writeJsonDataToFile(anyString(), anyString(), any(Timestamp.class), anyString())).thenReturn(new LogResponseModel());
+        PollDataSyncConfig config = new PollDataSyncConfig();
+        config.setTableName(tableName);
+        config.setKeyList("key");
+        config.setSourceDb("RDB");
         // Act
-        universalDataHandlingService.updateDataHelper(exceptionAtApiLevel, tableName, timestamp,
-                rawJsonData, isInitialLoad, log, "RDB", "key");
+        universalDataHandlingService.updateDataHelper(exceptionAtApiLevel, timestamp,
+                rawJsonData, isInitialLoad, log, timestamp, config);
 
         // Assert
         verify(iPollCommonService, times(1)).writeJsonDataToFile(RDB, tableName, timestamp, rawJsonData);
-        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLogLocalDir(tableName, timestamp, LOCAL_DIR_LOG + "Local File Save Success");
+        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLogLocalDir(
+                eq(tableName), eq(timestamp),
+                any());
     }
 
 
@@ -360,13 +320,17 @@ class RdbModernDataHandlingServiceTest {
 
         when(is3DataService.persistToS3MultiPart(anyString(), anyString(), anyString(), any(Timestamp.class), anyBoolean()))
                 .thenThrow(new RuntimeException("S3 Error"));
-
+        PollDataSyncConfig config = new PollDataSyncConfig();
+        config.setTableName(tableName);
+        config.setKeyList("key");
+        config.setSourceDb("RDB");
         // Act
-        universalDataHandlingService.updateDataHelper(exceptionAtApiLevel, tableName, timestamp,
-                rawJsonData, isInitialLoad, log, "RDB", "key");
+        universalDataHandlingService.updateDataHelper(exceptionAtApiLevel, timestamp,
+                rawJsonData, isInitialLoad, log, timestamp, config);
 
         // Assert
-        verify(iPollCommonService, times(1)).updateLastUpdatedTimeAndLogLocalDir(tableName, timestamp, CRITICAL_NON_NULL_LOG + "S3 Error");
+        verify(iPollCommonService, times(1))
+                .updateLogNoTimestamp(eq(tableName), any());
     }
 
 }
